refactor: move search-lexical into search-consolidated file
Ref eng-371
